+++ src/pages/forgot-password/forgot-password.tsx
@@ -2,7 +2,7 @@ import styles from './forgot-password.module.css'
 import { Link, useLocation, Navigate } from 'react-router-dom'
 
 import { useEffect, FC } from 'react'
-import ForgotPasswordForm from '../../components/forgot-password-form/forgot-password-form'
+import ForgotPasswordForm from '../../components/forms/forgot-password-form/forgot-password-form'
 import { SET_INIT_STATE_RESTORE_PASS } from '../../services/actions/restore-pass'
 import { useAppSelector, useAppDispatch } from '../../services/custom-hooks/redux-hooks'
