+++ src/pages/login/login.tsx
@@ -1,6 +1,6 @@
 import styles from './login.module.css'
 import { Link, Navigate, useLocation } from 'react-router-dom'
-import LoginForm from '../../components/login-form/login-form'
+import LoginForm from '../../components/forms/login-form/login-form'
 import { useAppSelector } from '../../services/custom-hooks/redux-hooks'
 import { FC } from 'react'
