+++ src/pages/reset-password/reset-password.tsx
@@ -1,7 +1,7 @@
 import styles from './reset-password.module.css'
 import { Link, Navigate, useLocation } from 'react-router-dom'
 import { useEffect, FC } from 'react'
-import ResetPasswordForm from '../../components/reset-password-form/reset-password-form'
+import ResetPasswordForm from '../../components/forms/reset-password-form/reset-password-form'
 import { SET_INIT_STATE_RESET_PASS } from '../../services/actions/reset-pass'
 import { useAppSelector, useAppDispatch } from '../../services/custom-hooks/redux-hooks'
